day13b: Solve second part
Instead of just stopping when we find a collision, remove the colliding
cars immediately.  We track this by giving each car an id, and
maintaining a set of ids that have been removed.  If we encounter one
that has already been removed (happens if the colliding car moves down
or to the right), we skip it.  Otherwise, we will remove all of the
removed ones before applying the final result.