day14: Solve both
The second part is a bit tricky because the pattern may match between
the two digits of a multiple digit answer.  Without this check, the
result is much too large.

A key is realizing that the sums are always of single digit numbers, and
will always either be a single digit, or a two digit number starting
with a 1.